divide up partnership page
